requirements.txt
openai>=1.3.0
gradio>=4.0.0
openai-whisper>=20231117
python-dotenv>=1.0.0
